+++ llama_index/chat_engine/condense_question.py
@@ -1,7 +1,11 @@
 import logging
 from typing import Any, List, Type, Optional
 
-from llama_index.chat_engine.types import BaseChatEngine, STREAMING_CHAT_RESPONSE_TYPE
+from llama_index.chat_engine.types import (
+    BaseChatEngine,
+    AgentChatResponse,
+    StreamingAgentChatResponse,
+)
 from llama_index.chat_engine.utils import response_gen_with_chat_history
 from llama_index.indices.query.base import BaseQueryEngine
 from llama_index.indices.service_context import ServiceContext
@@ -9,7 +13,8 @@ from llama_index.llms.base import ChatMessage, MessageRole
 from llama_index.llms.generic_utils import messages_to_history_str
 from llama_index.memory import BaseMemory, ChatMemoryBuffer
 from llama_index.prompts.base import Prompt
-from llama_index.response.schema import RESPONSE_TYPE, StreamingResponse
+from llama_index.response.schema import StreamingResponse, RESPONSE_TYPE
+from llama_index.tools import ToolOutput
 
 logger = logging.getLogger(__name__)
 
@@ -123,9 +128,19 @@ class CondenseQuestionChatEngine(BaseChatEngine):
         )
         return response
 
+    def _get_tool_output_from_response(
+        self, query: str, response: RESPONSE_TYPE
+    ) -> ToolOutput:
+        return ToolOutput(
+            content=str(response),
+            tool_name="query_engine",
+            raw_input={"query": query},
+            raw_output=response,
+        )
+
     def chat(
         self, message: str, chat_history: Optional[List[ChatMessage]] = None
-    ) -> RESPONSE_TYPE:
+    ) -> AgentChatResponse:
         chat_history = chat_history or self._memory.get()
 
         # Generate standalone question from conversation context and last message
@@ -137,17 +152,22 @@ class CondenseQuestionChatEngine(BaseChatEngine):
             print(log_str)
 
         # Query with standalone question
-        response = self._query_engine.query(condensed_question)
+        query_response = self._query_engine.query(condensed_question)
+        tool_output = self._get_tool_output_from_response(
+            condensed_question, query_response
+        )
 
         # Record response
         self._memory.put(ChatMessage(role=MessageRole.USER, content=message))
-        self._memory.put(ChatMessage(role=MessageRole.ASSISTANT, content=str(response)))
+        self._memory.put(
+            ChatMessage(role=MessageRole.ASSISTANT, content=str(query_response))
+        )
 
-        return response
+        return AgentChatResponse(response=str(query_response), sources=[tool_output])
 
     def stream_chat(
         self, message: str, chat_history: Optional[List[ChatMessage]] = None
-    ) -> STREAMING_CHAT_RESPONSE_TYPE:
+    ) -> StreamingAgentChatResponse:
         chat_history = chat_history or self._memory.get()
 
         # Generate standalone question from conversation context and last message
@@ -159,20 +179,22 @@ class CondenseQuestionChatEngine(BaseChatEngine):
             print(log_str)
 
         # Query with standalone question
-        response = self._query_engine.query(condensed_question)
+        query_response = self._query_engine.query(condensed_question)
+        tool_output = self._get_tool_output_from_response(
+            condensed_question, query_response
+        )
 
         # Record response
         if (
-            isinstance(response, StreamingResponse)
-            and response.response_gen is not None
+            isinstance(query_response, StreamingResponse)
+            and query_response.response_gen is not None
         ):
             # override the generator to include writing to chat history
-            response = StreamingResponse(
-                response_gen_with_chat_history(
-                    message, self._memory, response.response_gen
+            response = StreamingAgentChatResponse(
+                chat_stream=response_gen_with_chat_history(
+                    message, self._memory, query_response.response_gen
                 ),
-                source_nodes=response.source_nodes,
-                metadata=response.metadata,
+                sources=[tool_output],
             )
         else:
             raise ValueError("Streaming is not enabled. Please use chat() instead.")
@@ -180,7 +202,7 @@ class CondenseQuestionChatEngine(BaseChatEngine):
 
     async def achat(
         self, message: str, chat_history: Optional[List[ChatMessage]] = None
-    ) -> RESPONSE_TYPE:
+    ) -> AgentChatResponse:
         chat_history = chat_history or self._memory.get()
 
         # Generate standalone question from conversation context and last message
@@ -192,17 +214,22 @@ class CondenseQuestionChatEngine(BaseChatEngine):
             print(log_str)
 
         # Query with standalone question
-        response = await self._query_engine.aquery(condensed_question)
+        query_response = await self._query_engine.aquery(condensed_question)
+        tool_output = self._get_tool_output_from_response(
+            condensed_question, query_response
+        )
 
         # Record response
         self._memory.put(ChatMessage(role=MessageRole.USER, content=message))
-        self._memory.put(ChatMessage(role=MessageRole.ASSISTANT, content=str(response)))
+        self._memory.put(
+            ChatMessage(role=MessageRole.ASSISTANT, content=str(query_response))
+        )
 
-        return response
+        return AgentChatResponse(response=str(query_response), sources=[tool_output])
 
     async def astream_chat(
         self, message: str, chat_history: Optional[List[ChatMessage]] = None
-    ) -> STREAMING_CHAT_RESPONSE_TYPE:
+    ) -> StreamingAgentChatResponse:
         chat_history = chat_history or self._memory.get()
 
         # Generate standalone question from conversation context and last message
@@ -214,20 +241,23 @@ class CondenseQuestionChatEngine(BaseChatEngine):
             print(log_str)
 
         # Query with standalone question
-        response = await self._query_engine.aquery(condensed_question)
+        query_response = await self._query_engine.aquery(condensed_question)
+        tool_output = self._get_tool_output_from_response(
+            condensed_question, query_response
+        )
 
         # Record response
         if (
-            isinstance(response, StreamingResponse)
-            and response.response_gen is not None
+            isinstance(query_response, StreamingResponse)
+            and query_response.response_gen is not None
         ):
             # override the generator to include writing to chat history
-            response = StreamingResponse(
-                response_gen_with_chat_history(
-                    message, self._memory, response.response_gen
+            # TODO: query engine does not support async generator yet
+            response = StreamingAgentChatResponse(
+                chat_stream=response_gen_with_chat_history(
+                    message, self._memory, query_response.response_gen
                 ),
-                source_nodes=response.source_nodes,
-                metadata=response.metadata,
+                sources=[tool_output],
             )
         else:
             raise ValueError("Streaming is not enabled. Please use achat() instead.")