+++ tests/agent/react/test_react_agent.py
@@ -1,0 +1,95 @@
+from typing import Any, List, Sequence
+import pytest
+from llama_index.agent.react.base import ReActAgent
+from llama_index.llms.base import ChatMessage, ChatResponse, MessageRole
+from llama_index.llms.mock import MockLLM
+from llama_index.tools.function_tool import FunctionTool
+from llama_index.chat_engine.types import (
+    AgentChatResponse,
+)
+
+
+@pytest.fixture
+def add_tool() -> FunctionTool:
+    def add(a: int, b: int) -> int:
+        """Add two integers and returns the result integer."""
+        return a + b
+
+    return FunctionTool.from_defaults(fn=add)
+
+
+class MockChatLLM(MockLLM):
+    def __init__(self, responses: List[ChatMessage]) -> None:
+        self._i = 0  # call counter, determines which response to return
+        self._responses = responses  # list of responses to return
+
+    def chat(self, messages: Sequence[ChatMessage], **kwargs: Any) -> ChatResponse:
+        del messages  # unused
+        response = ChatResponse(
+            message=self._responses[self._i],
+        )
+        self._i += 1
+        return response
+
+
+MOCK_ACTION_RESPONSE = """\
+Thought: I need to use a tool to help me answer the question.
+Action: add
+Action Input: {"a": 1, "b": 1}
+"""
+
+MOCK_FINAL_RESPONSE = """\
+Thought: I have enough information to answer the question without using any more tools.
+Answer: 2
+"""
+
+
+def test_chat_basic(
+    add_tool: FunctionTool,
+) -> None:
+    mock_llm = MockChatLLM(
+        responses=[
+            ChatMessage(
+                content=MOCK_ACTION_RESPONSE,
+                role=MessageRole.ASSISTANT,
+            ),
+            ChatMessage(
+                content=MOCK_FINAL_RESPONSE,
+                role=MessageRole.ASSISTANT,
+            ),
+        ]
+    )
+
+    agent = ReActAgent.from_tools(
+        tools=[add_tool],
+        llm=mock_llm,
+    )
+    response = agent.chat("What is 1 + 1?")
+    assert isinstance(response, AgentChatResponse)
+    assert response.response == "2"
+
+
+@pytest.mark.asyncio
+async def test_achat_basic(
+    add_tool: FunctionTool,
+) -> None:
+    mock_llm = MockChatLLM(
+        responses=[
+            ChatMessage(
+                content=MOCK_ACTION_RESPONSE,
+                role=MessageRole.ASSISTANT,
+            ),
+            ChatMessage(
+                content=MOCK_FINAL_RESPONSE,
+                role=MessageRole.ASSISTANT,
+            ),
+        ]
+    )
+
+    agent = ReActAgent.from_tools(
+        tools=[add_tool],
+        llm=mock_llm,
+    )
+    response = await agent.achat("What is 1 + 1?")
+    assert isinstance(response, AgentChatResponse)
+    assert response.response == "2"