+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 ## Unreleased
 
 ### New Features
+- Added sources to agent/chat engine responses (#6854)
 - Added basic chat buffer memory to agents / chat engines (#6857)
+
+### Bug Fixes / Nits
+- Fixed support for legacy vector store metadata (#6867)
 
 ## [v0.7.5] - 2023-07-11